scripts/requirements.txt
python-dotenv
spotify